requirements.txt
psycopg2-binary
python-dotenv~=1.0.1
chromadb
streamlit
streamlit_extras

pymongo~=4.8.0
psycopg2~=2.9.9
pandas~=2.2.2
openai
pdfplumber~=0.11.4